+++ e2e/gateway_e2e_test.go
@@ -34,7 +34,7 @@ func requireAPIKey(t *testing.T) string {
 }
 
 func gatewayBaseURL() string {
-	return getEnv("GATEWAY_BASE_URL", "http://127.0.0.1:8080")
+	return getEnv("GATEWAY_BASE_URL", "http://127.0.0.1:6001")
 }
 
 func httpClient() *http.Client {